+++ ios/Pods/Target Support Files/Pods-CalendarsExample/Pods-CalendarsExample-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_CalendarsExample : NSObject
-@end
-@implementation PodsDummy_Pods_CalendarsExample
-@end