+++ CHANGELOG.md
@@ -24,8 +24,8 @@
 
 ## [1.192.0] - 2019-06-25
 ### Fix
-- CalendarProvider - Fix for warning on Image 'source' type.
-- ExpandableCalendar - Fix shadow/elevation style.
+- CalendarProvider - fix for warning on Image 'source' type.
+- ExpandableCalendar - fix shadow/elevation style.
 
 ## [1.193.0] - 2019-06-26
 ### Bug Fix
@@ -99,4 +99,49 @@
 
 ## [1.209.0] - 2019-08-20
 ### Fix
-- CalendarProvider - Fix for children wrapper view.
+- CalendarProvider - fix for children wrapper view.
+
+## [1.213.0] - 2019-11-13
+### Changed
+- Upgrading react-native to version 61.4.0.
+
+## [1.214.0] - 2019-12-12
+### Fix
+- Agenda - fix typo on 'scrollEnable' prop.
+- ExpandableCalendar - fix passed props to CalendarList.
+
+## [1.215.0] - 2019-12-23
+### Added
+- ExpandableCalendar - adding week scroll to closed position.
+
+## [1.216.0] - 2019-12-24
+### Fix
+- Week - moving width style to style file.
+
+## [1.217.0] - 2019-12-24
+### Fix
+- Week - moving width style to WeekCalendar.
+
+## [1.218.0] - 2020-1-12
+### Added
+- AgendaList - allow text transform override for section header text.
+
+## [1.221.0] - 2020-1-16
+### Fix 
+- ExpandableCalendar - fix renderArrow function (PR #1037).
+
+## [1.223.0] - 2020-1-16
+### Added
+- WeekCalendar - new component.
+
+## [1.224.0] - 2020-1-16
+### Added
+- AgendaList - allow passing keyExtractor.
+
+## [1.225.0] - 2020-1-19
+### Fix
+- package.json - adding hoist-non-react-statics dependency (fix issue #900).
+
+## [1.226.0] - 2020-1-19
+### Added
+- Life cycle methods - renaming with UNSAFE_ prefix componentWillMount, componentWillReceiveProps.